--- client/src/views/auth/login/login.css
@@ -1,44 +1,0 @@
-
-.login-container {
-    display: flex;
-    flex-direction: column;
-    height: 100%;
-    font-family: 'Poppins',sans-serif;
-    
-  }
-  
-  .login-page {
-    flex: 1;
-  }
-
-  .Right-panel{
-    justify-content:center;
-    align-items:center;
-    border-radius: 15px;
-  }
-
-  .slide-enter {
-    transform: translateX(-100%);
-  }
-  
-  .slide-enter-active {
-    transform: translateX(0);
-    transition: transform 500ms;
-  }
-  
-  .slide-exit {
-    transform: translateX(0);
-  }
-  
-  .slide-exit-active {
-    transform: translateX(-100%);
-    transition: transform 500ms;
-  }
-  
-  .slide-image-container {
-    position: relative;
-  }
-  .login-section{
-    margin:0 4rem;
-    border-radius:15px;
-  }
